requirements.txt
openai
PyPDF2
python-docx
streamlit
python-dotenv